+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [21.1.0] - 2024-11-19
+
+-   Adds `getCookieNameForTokenType` config option to allow customizing the cookie name for a token type.
+-   Adds `getResponseHeaderNameForTokenType` config option to allow customizing the response header name for a token type.
+    -   Please note, that using this will require further customizations on the frontend
+
 ## [21.0.0] - 2024-10-07
 
 -   Added OAuth2Provider recipe